Add AST to trace instructions. Sync some with master.

+++ interp/interp.go
@@ -541,7 +541,7 @@ func runFrame(fr *Frame) {
 		}
 	block:
 		// rocky: changed to allow for debugger "jump" command
-		for fr.pc = 0; fr.pc < uint(len(fr.block.Instrs)); fr.pc++ {
+		for fr.pc = 0; fr.pc < len(fr.block.Instrs); fr.pc++ {
 			instr = fr.block.Instrs[fr.pc]
 			if InstTracing() {
 				fmt.Fprint(os.Stderr, fr.block.Index, fr.pc, "\t")
